fix race condition and remove lazy_static
Based on feedback (thank-you!), the timer in the into_panic test has
been replaced with `futures::future::pending()` and the lazy_static
dev-dependency has been removed and replaced with `const_mutex` from
`parking_lot`.

Additionally, the panic catching has been reworked a little bit to set
and unset the panic hook around the part of the test where the panic is
expected. This fixes the issue that the normal test output for a failed
test was being eaten up by the custom panic hook.