doc: update clippy command in contribution guide
The contribution guide describes the Tokio code of conduct and gives a
list of steps to follow when contributing in different ways. In the
guide to contributing a pull request, commands are listed to be run by
the contributor locally to help ensure that the PR will pass on CI.

The command to run clippy isn't the same as the one run on CI,
specifically it the command doesn't check the tests. This commit changes
the command to match the one on CI, so that tests are checked by clippy
as well.