rt: add rng_seed option to `runtime::Builder`
The `tokio::select!` macro polls branches in a random order. While this
is desirable in production, for testing purposes a more deterministic
approach can be useul.

This change adds an additional parameter to the runtime `Builder` to set
the random number generator seed. This value is then used to reset the
seed on all threads associated with the runtime being built.

This guarantees that calls to the `tokio::select!` macro which are
performed in the same order on the same thread will poll branches in the
same order.