builder accepts RngSeed opaque struct
NOTE: This change doesn't correct the handling of the RngSeed on the
thread the runtime is started from, so it's all likely to change. Just
pushing for safety.

Instead of exposing the width of the seed we're currently using, expose
an opaque struct which can generate a seed from a byte slice.

Additionally, each thread is given with a unique seed based on its `id`
(`worker_thread_index`). This should enable more deterministic behavior,
while ensuring that each thread does not begin with the same seed.