rt: move I/O driver into `runtime` module
This patch moves the I/O driver into the runtime module. The I/O driver
is a runtime concern and is only used by the runtime. Moving the driver
is a first step to clean up some Tokio internals. There will be follow
up patches that integrate the I/O driver and other runtime concerns more
closely.

This is an internal refactor and should not impact any public APIs.