uring: include `tokio-uring` as a module
The `tokio-uring` crate is placed as a module  at
`tokio::platform::linux::uring` and all module paths inside have been
updated to match the new path. The 'tokio_uring::runtime` module is
removed and io_uring will be integrated with the `tokio::runtime` in
future commits. The inclusion of the module in the module tree is
conditioned on the 'uring' feature flag.

io_uring object and its operations slab are protected by a mutex; this
can be improved in the future by creating a thread-per-core runtime or
using a sharded mutex for the multi-threaded runtime.

Other changes have been made to the driver to prepare it for
integration into the `tokio` runtime.