`is_closed` for `mpsc::{Receiver, UnboundedReceiver}`
There's a use case for channels that wants to know up-front (before
calling `recv` or `try_recv`)that there's a possibility of new messages
coming in. Make it possible to check whether the channel is closed and
empty, which implies that there will never be new messages on this
channel.

Fixes: #4638