tokio_util: add inspection wrapper for `AsyncWrite`
When writing things out, it's useful to be able to inspect the bytes
that are being written and do things like hash them as they go past.
This isn't trivial to get right, due to partial writes and efficiently
handling vectored writes (if used).

Provide an `InspectWriter` wrapper that gets this right, giving a
supplied `FnMut` closure a chance to inspect the buffers that have been
successfully written out.

Fixes: #4584