macros: rename `thread_local!` to `tokio_thread_local!`
## Motivation

Currently, Tokio uses an internal `thread_local!` macro which compiles
to either `std::thread_local!` or `loom::thread_local!`, and switches
between using `const` expressions in thread locals or not based on cfg
attributes. Because this macro has the same name as
`std::thread_local!`, though, which is ambiently available, it's not
always clear which macro is being used, and it seems like there are
sometimes name resolution issues that can result in the redefined macro
not being used as we'd expect.

## Solution

This branch renames the internal macro to `tokio_thread_local!` to make
it explicit that we want to use it, rather than `std::thread_local!`.
I've removed some `cfg` attributes in `local_set.rs` that shouldn't be
necessary, since our macro *should* handle that already.