rt: Resolve panic with large duration sleeps
Durations far into the future would cause the conversion of millis
from u128 to u64 to fail and set the value to u64::MAX. This in turn
causes issues since u64::MAX is used as a signal value.

This patch adds a new constant that is the largest non-signal value
that can be used for ticks. And uses that in the location it causes
issues.