Draft: Add BWoS-queue backend to tokio
This commit is a work-in-progress snapshot of BWoS for
tokio, with the intention to get early feedback.
Currently, the BWoS queue is just dropped in as a replacement,
with the intention to make benchmarking easier (just patch
downstream crates to use the modified version).

Before merging the queue should be integrated as an alternate
queue instead of replacing the current one.

The design of the BWoS queue was done by Jiawei Wang.

Co-authored-by: Jiawei Wang <jiawei.wang@huawei.com>
Designed-by: Jiawei Wang <jiawei.wang@huawei.com>
Signed-off-by: Jonathan Schwender <jonathan.schwender@huawei.com>
Signed-off-by: Jiawei Wang <jiawei.wang@huawei.com>
Signed-off-by: Ming Fu <ming.fu@huawei.com>