task: refactor: rename State::has_join_waker to State::is_join_waker_set
The JOIN_WAKER bit in State is unlike the other bits (e.g. RUNNING, COMPLETE),
which indicate a binary condition. The JOIN_WAKER bit is part of a protocol that
enables safe access to the waker field of a task (described in task/mod.rs).
Although currently it so happens that if JOIN_WAKER is set, then there is a
waker in the waker field, this is incidental and can change in the future.
Rename has_join_waker to a more neutral is_join_waker_set to decouple the
meaning of the bit being set from having a waker in the waker field.