notify: make sequential notify_one calls behave like a single call.
According to the docs `notify_one()` sets a permit if there currently
are no available permits and the permit should be consumed only by
`notified().await`.
Before this commit, if only a single `notified().await` was called
before `notify_one()`, then `notify_one()` consumed its own permit by
setting the state to EMPTY. A subsequent call to `notify_one()` would
then set a new permit even though the previous permit was not yet
consumed by the `notified().await`.

This commit solves the issue by transferring the responsibility of
setting the state to EMPTY to the notified instead of the notifier.

Co-authored-by: Shachar Langbeheim <nihohit@gmail.com>