rt: more surgical check for awoken
Lots of tests seemed to require that their awoke tasks would run after a
park. So be surgical about the new check. Only do the park if the new
tasks set is empty and there were awoken before the `before_park`
callback er even after `before_park` there still aren't any awoken.

This is meant to skip the park when the `before_park` has awoken one or
more tasks but we can't afford to skip the park if there had already
been awoken beforehand. Seems that according to the tests, time doesn't
march foreward in that case, or more accurately, timers don't expire in
that case.